Always append one double quote char in the end of string
The returned value write_string_or_name() is not really checked, so one double quote char error
is not identified and the broken json string is directly sent to the client.

To address the issue, the workaround is always appending one double quote char in the end of string.

TODO:
1. Find the root cause
2. Check the returned value of write_string_or_name()

Longhorn 6190

Signed-off-by: Derek Su <derek.su@suse.com>