blob: add blob_set_back_bs_dev function
This function does mostly the same as what spdk_blob_set_esnap_bs_dev
did before, with a new optional callback to be called after the
destroy of the old back_bs_dev and before the set of the new one.
This callback will be used in reparent operations to change the
references of the blob's parent.

Signed-off-by: Damiano Cipriani <damiano.cipriani@suse.com>